438 createAsyncThunk
zapochvame da fetchvame dannite ot link, vmesto ot fayl v prilozhenieto.
We cannot just simoly set this up in our current reducers, it's not going to work.
That's why with redux toolkit, we install another library, the think one, and
from the redux toolkit we get this createAsyncThunk and we right away wanna
invoke it and we wanna export the result